Fixed Collision with Xander and Cat objects, made a folder of Test Scenes, made Enemy Script
Hey guys, for this I fixed the collision between the Xander and Cat objects (I did this by setting the Rigidbody components on both of these to Dynamic instead of Kinematic). In addition to this, I added movement and more alert logic to the Enemy Script. Also, I made a folder of test scenes as well, where each person has their own scene. It is vital that everyone uses their own test scene when testing things, because if multiple people modify the same scene it will create merge conflicts!!!