Did added work on Grandmother Object, making the Grandmother attack with the Frying pan
In this Update, I further added to the Enemy Script, and then I also added more to the Grandmother script. I created the Grandmother Prefab object, and the grandmother can also hit Xander with the Frying Pan when she is close enough to him. For know, the Grandmother uses the same proto Skeleton assets I found, and it uses the same animation from the Enemy base object prefab. In addition to this, adding additional logic, such as taking health from Xander, will need to be added to the Damage scripts OnCollisionEnter2D Method.