Planned out how to update the Movement System with the Grandmother
In this commit, I mainly just planned on how to revamp the movement system for the Grandmother (and Cat), using Waypoints. Specifically, I am planning on making it so a waypoint object will have a list of potential waypoints to travel to, and then the Grandmother will decide from those which one to travel to. This will seek to avoid having the Grandmother or Cat move into objects that have collision, such as tables. I also planned how the Grandmother or Cat would exit the room with this system.