Created Waypoint, EnemyRoomRoaming scripts, and created initial version of new Movement system for Grandmother and Cat
Began working on the initial version of the Movement pattern script. I created a new script called EnemyRoomRoaming, which will handle the new Waypoint movement system for the Cat and Grandmother objects. Essentially, these objects will begin at a starting node in a room, and then travel randomly to Waypoints. The next waypoint that they travel to will be determined by the current waypoint the object is at.
This is achieved with every Waypoint object now having a Waypoint script attached to it, which will have a list of potential nodes to travel to. In addition to this, if the Grandmother must leave the room (after the grandmother hits a certain number of nodes, the grandmother will leave the room), then each Waypoint will also have a reference to the next node on the path to the exit. NOTE, all of these Waypoints, in terms of the available waypoints and path to exit node, MUST be set in the Inspector with how I am designing it. This means that it can also be easily reusable, because you would just plop the grandmother object in the scene, and then if the grandmother is currently supposed to be active, make it active at the beginning doorway and then it can travel to a certain amount of nodes and then leave.