Implemented the ability for the EnemyRoomRoaming enemies to move in between scenes successfully
In this update, I implemented the enemies who are of type EnemyRoomRoaming (Grandmother) to be able to successfully "travel" in between scenes. It does this by having a static currentRoomIndex, which all of the Grandmother objects have. So, if the grandmother is in the current room you are in, and then she travels to the next room in the sequence, if you travel to that room as well, she will be in that room. In addition to this, I also added functionality for the Grandmother (EnemyRoomRoaming class) to tell which exit (North, South, East, or West) they should go to, depending on the next room they need to go to.