Put Grandmother and Cat Enemies into Alex 1,2, and 3 Scenes
In this commit, I added the EnemyRoomRoaming Functionality for the Grandmother and Cat enemies into the Alex1, 2, and 3 scenes. For this, I made the RoomAlex1, 2, 3, and 4 prefabs, and linked them together based on the scene name. Thus, it will be VERY important to make sure that we just use the Alex 1, 2, and 3 scene names for the main levels for the Alpha Build. Also, I inserted the Waypoints, and linked them up appropriately as well. All of the needed  Enemy Roaming Functionality for the Grandmother and Cat enemies for the alpha build should be complete now. The only things that would be needed would be inserting the new Grandmother and Cat sprites whenever they are ready