Added and Modified Enemy's in new Bedroom1 and Kitchen Scenes
In this commit, I modified the Grandmother and Cat enemies to allow them to properly function in the new Bedroom1 and Kitchen scenes. In doing this, I had to revamp the enemy movement system when we transitioned from Alex's scenes to these ones. In addition to this, I also was able to fix a but that was causing the Grandmother to stop traveling to nodes after she woke up from being "asleep". However, I did notice some errors in editor with some other people's work, when you run the Bedroom1 scene you will get these errors:

"NullReferenceException: Object reference not set to an instance of an object
ItemSpawner.Start () (at Assets/Scripts/ItemSpawner.cs:29)"

NullReferenceException: Object reference not set to an instance of an object
ItemSpawner.Start () (at Assets/Scripts/ItemSpawner.cs:29)