Added Waypoints to All of the Rooms
In this Commit, I added the Waypoints to all of the Scenes (with the Exception of the Upstairs Bathroom, I am planning on just not having waypoints in there). I also created the Room Objects to "daisy-chain" the Rooms together for the EnemyRoomRoaming objects (Cat and Grandmother). I was researching the A* Algorithm to make the Roaming enemy find the path to the exit, but I stopped and I just put in the nearest node to each exit manually in the Inspector. Then, if I have extra time later either during this phase or during the Gold Phase, I can revisit it. In addition to this, the Scenes were added to the Build Settings, and I also fixed the Collision Bounds to the majority of the scenes, I think they might still need to be fixed in some of the scenes. In addition to this, I added it to most of them, but in Alex's scenes, you will just want to make sure that the Main Camera object from the other scenes (the one with the additional scripts, and the Map Bounds child object) is in that scene instead of the default one that is already in that scene.