Implemented the Scene Transition for the house rooms and the Item system in the rooms
Added the scene transition and item scripts to all determined room layouts in the house.  As of right now, all transitions points are correctly connected with each other and are working. There is only a few current transition points that have a current bug where the scene transition does not perform immediately. Also added the Item spawned and spawn points, Xander character in every main room. In the Bathroom, 2F Hallway, and Bottom Left Bedroom the layers will need to be reorder since currently the tiles cover the player and items. Currently when testing it seems the damage from bad items, death cheese, is now currently working. Also added the rest of the improved tiles into the project and our under the folder named updated tiles in the map tiles folder under the alpha dog arts folder. Replaced most of the single tile door frames with new tiles except in the same three rooms as the sprite layout order. Got the pause feature now established where the game is fully paused and un-paused from the pause menu. The options menu UI does currently have a bug issue which I think is from the Unity Editor. Also the UI and pause feature are only set to be possible in non-menu scenes. Recommend suggestions that we should extend the upstairs hallway, bathroom and bottom left bedroom by a few tile rows/columns.