Tested Grandmother and Cat Enemies, and Modified Scenes
In this commit, I did substantial work in the Build Scenes. Firstly, I modified the MapCollision objects on the Main Camera, to make sure that each of them were scaled approprietly with the map they were currently in. In addition ot that, I also modified the Boundary collisions of every level to make sure that they fit the level, and were properly updated. Then, I also moved around certain waypoints to ensure that they accommodated tile changes that occured since Saturday night (when I first put the waypoints in there). Finally, I also tested the Grandmother and Cat enemies, making sure that they traveled to each of the rooms correctly. I also made the Grandmother's home in the UpstairsTopLeftBedroom scene, and I made the Cat's Home Room in the DownstairsBottomRightLivingRoom scene, and I added a placeholder Cat Bed to accommodate this. I also put the Grandmother and Cat objects in the first scene loaded (DownstairsTopLeftKitchen) to make sure that both of them were created at the beginning of the game. The Cat will begin in the Kitchen like we discussed in our meeting, and to do this I hardcoded the cat to be in that room first, and then go back to it's Home room, or the DownstairsBottomRightLivingRoom, to go to sleep.