Wolf Spider Functionality Major Update
In this Commit, I did major work on the Wolf Spider Roaming Script and Top-Down Wolf Spider. Here, the WolfSpiderRoaming Script is practically complete, with the exception of bug testing of course. In addition to this, it seems like it has also been able to successfully spawn Side and Top-Down Wolf Spiders successfully, and also signal for them to travel to the exit as well. In addition to this, I also upgraded the Room and RoomManager scripts. Firstly, the Room object now will pull information from the PerspectiveSceneChange objects in the scene to see where the SpawnPoints are and Exits are in the scene. Essentially, if the PerspectiveSceneChange objects are in the Climbing Levels, all we will need to do is create a Room object in that scene, and click isWallCrawlingScene in the Hierarchy, and the rest of the fields (sceneName, Exits, and scenes that Room is linked too) will be autogenerated by the class, from using the information that is in the PerspectiveSceneChange objects.