Updated EnemyWaypoints
In this commit, I updated the WolfSpiderRoaming, Room, and EnemyWaypoint Scripts. Essentially, I added another bool to the EnemyWaypoints called isSpawnNode, and then the Room script would make a list of waypoints that are labeled as spawn nodes using the variable mentioned above. Then, the WolfSpider will only spawn at one of those nodes if Xander is entering a room that the Spider is currently in (for the Top-Down one at least). In addition to this, I went through all of the Top-Down scenes and adjusted the waypoints, and then also labeled nodes as spawnNodes, to make it so the spider would not spawn right in front of Xander when he is entering a room. However, I did not quite get a chance to extensively test this, and I have to still also implement this into the EnemyRoomRoaming functionality as well.