Worked more on adding Enemy Collision Avoiding Logic
In this commit, I worked more on the enemy avoiding logic that we had discussed about adding. I moved the logic for this from Cat.cs to the Enemy.cs base class. In addition to this, I seem to have fixed the issue of the "stuck" variable firing (the logic for the Cat being stuck) when it is not supposed to. Instead of using the distanceTowardsXander, which I was using before, I am now using Raycasts to determine if the enemy is blocked by an obstacle. It is not necessarily perfect, and I am noticing a few bugs, but it is better than the solution I had previously had. I will show you guys my progress on this tomorrow night in the meeting, and we can discuss on how you guys feel about it.